refactor: :truck: rename verifiers files and classes
