+++ scripts/verification/verifier_etherscan.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 """
-Sourcify block explorer verifier implementation using forge verify-contract.
+Etherscan block explorer verifier implementation using forge verify-contract.
 """
 import sys
 import subprocess
@@ -10,19 +10,21 @@ from .retry import retry_with_backoff
 # Block explorer configurations
 CHAIN_ID = '1'  # Mainnet only
 
-class SourcifyVerifier:
-    """Sourcify block explorer verifier using forge verify-contract."""
+
+class VerifierEtherscan:
+    """Etherscan block explorer verifier using forge verify-contract."""
     
-    def __init__(self, chain_id: str):
+    def __init__(self, api_key: str, chain_id: str):
+        self.api_key = api_key
         self.chain_id = chain_id
     
     def is_available(self) -> bool:
-        """Check if Sourcify supports this chain."""
+        """Check if Etherscan supports this chain."""
         return self.chain_id == CHAIN_ID
     
     def get_verification_url(self, contract_address: str) -> str:
-        """Get Sourcify URL for the verified contract."""
-        return f"https://sourcify.dev/#/lookup/{contract_address}"
+        """Get Etherscan URL for the verified contract."""
+        return f"https://etherscan.io/address/{contract_address}#code"
     
     @retry_with_backoff(max_retries=3, base_delay=2, max_delay=30)
     def verify_contract(
@@ -32,15 +34,16 @@ class SourcifyVerifier:
         constructor_args: str,
         library_address: str = ""
     ) -> bool:
-        """Verify contract on Sourcify using forge verify-contract."""
-        print(f'\nVerifying {contract_name} at {contract_address} on Sourcify...')
+        """Verify contract on Etherscan using forge verify-contract."""
+        print(f'\nVerifying {contract_name} at {contract_address} on Etherscan...')
         
         # Build forge verify-contract command
         cmd = [
             'forge', 'verify-contract',
             contract_address,
             f'src/DssSpell.sol:{contract_name}',
-            '--verifier', 'sourcify',
+            '--verifier', 'etherscan',
+            '--etherscan-api-key', self.api_key,
             '--flatten',
             '--watch'
         ]
@@ -67,7 +70,7 @@ class SourcifyVerifier:
         except subprocess.CalledProcessError as e:
             # Check if it's already verified
             if 'already verified' in e.stderr.lower() or 'already verified' in e.stdout.lower():
-                print('✓ Contract is already verified on Sourcify')
+                print('✓ Contract is already verified on Etherscan')
                 return True
             
             print(f"✗ Verification failed: {e.stderr}", file=sys.stderr)
